Rename system_domain_worker.rs to domain_worker_starter.rs
