+++ domains/client/domain-executor/src/lib.rs
@@ -87,13 +87,13 @@ mod domain_block_processor;
 mod domain_bundle_producer;
 mod domain_bundle_proposer;
 mod domain_worker;
+mod domain_worker_starter;
 mod executor;
 mod fraud_proof;
 mod gossip_message_validator;
 mod parent_chain;
 mod sortition;
 mod system_bundle_processor;
-mod system_domain_worker;
 mod system_gossip_message_validator;
 // TODO: Unlock once domain test infra is workable again.
 // #[cfg(test)]
